user and camera repositories, controllers

+++ server/controllers/camera.controller.js
@@ -1,21 +1,17 @@
 import cameraRepository from '../repositories/camera.repository.js';
 
-// import { getDirsPaths, getDirsNames } from '../utils/camera.utils.js';
-// import { makeDir, writeFile, removeDir } from '../services/fileService.js';
-
-// console.log('cameraController');
-
 const getAll = async ({ userId }) => {
-  // console.log('- camera controller getAll -', { userId });
   return await cameraRepository.getAll({ userId });
 };
 
 const getOne = async ({ userId, cameraId }) => {
   // console.log('- camera controller getOne -', { userId, cameraId });
   const camera = await cameraRepository.getOne({ userId, cameraId });
+
   if (!camera) {
     throw new Error('camera not found');
   }
+
   return camera;
 };
 
@@ -26,12 +22,26 @@ const createOne = async ({ userId, payload }) => {
 
 const updateOne = async ({ userId, cameraId, payload }) => {
   // console.log('- camera controller updateOne -', { userId, cameraId, payload });
-  return await cameraRepository.updateOne({ userId, cameraId, payload });
+  const camera = await cameraRepository.getOne({ userId, cameraId });
+
+  if (!camera) {
+    throw new Error('camera not found');
+  }
+
+  await cameraRepository.updateOne({ userId, cameraId, payload });
+
+  return await cameraRepository.getOne({ userId, cameraId });
 };
 
 const deleteOne = async ({ userId, cameraId }) => {
   // console.log('- camera controller deleteOne -', { userId, cameraId });
+  const camera = await cameraRepository.getOne({ userId, cameraId });
+
+  if (!camera) {
+    throw new Error('camera not found');
+  }
+
   return await cameraRepository.deleteOne({ userId, cameraId });
 };
 
-export default { getAll, createOne, getOne, updateOne, deleteOne };
+export default { getAll, getOne, createOne, updateOne, deleteOne };
